+++ Clubber.Discord/Modules/UserManagementCommands.cs
@@ -195,7 +195,7 @@ public sealed class UserManagementCommands(
                 return;
             }
 
-            uint leaderboardId = (uint)ddUser.LeaderboardId;
+            uint leaderboardId = ddUser.LeaderboardId;
             Task<IReadOnlyList<EntryResponse>> playerEntryTask = webService.GetLbPlayers([leaderboardId]);
             Task<GetPlayerHistory?> playerHistoryTask = webService.GetPlayerHistory(leaderboardId);
             await Task.WhenAll(playerEntryTask, playerHistoryTask);
